refactor(ui): extract SandboxedHtmlIframe shared component
The iframe render+resize+theme-sync pattern was duplicated in three files
(run-results-panel, agent-output-display, task-context-panel). Extract a
shared SandboxedHtmlIframe component alongside iframe-helpers and
deduplicate all three consumers. Also tighten isHtmlString to reject
strings like `<html-foo` that aren't full HTML documents.
