+++ packages/platform-ui/src/components/processes/run-results-panel.tsx
@@ -2,11 +2,10 @@
 
 import * as React from 'react';
 import { CheckCircle2, ExternalLink, Gauge, GitBranch, Clock, FileText, DollarSign } from 'lucide-react';
-import { useTheme } from 'next-themes';
 import type { StepExecution, AgentOutputSnapshot } from '@mediforce/platform-core';
 import { cn, isBrowsableRepoUrl } from '@/lib/utils';
 import { formatDuration, formatStepName, formatCostUsd } from '@/lib/format';
-import { buildSrcdoc, clampIframeHeight, isIframeResizeMessage } from '@/components/tasks/iframe-helpers';
+import { SandboxedHtmlIframe } from '@/components/tasks/sandboxed-html-iframe';
 
 interface RunResultsPanelProps {
   stepExecutions: StepExecution[];
@@ -49,44 +48,10 @@ function isEmptyResultValue(value: unknown): boolean {
 
 function isHtmlString(value: string): boolean {
   const trimmed = value.trimStart().toLowerCase();
-  return trimmed.startsWith('<!doctype html') || trimmed.startsWith('<html');
-}
-
-function HtmlPreview({ html }: { html: string }) {
-  const iframeRef = React.useRef<HTMLIFrameElement>(null);
-  const [height, setHeight] = React.useState(300);
-  const { resolvedTheme } = useTheme();
-  const isDark = resolvedTheme === 'dark';
-
-  React.useEffect(() => {
-    const handler = (event: MessageEvent) => {
-      if (
-        isIframeResizeMessage(event.data) &&
-        iframeRef.current &&
-        event.source === iframeRef.current.contentWindow
-      ) {
-        setHeight((prev) => {
-          const next = clampIframeHeight(event.data.height);
-          return next > 0 ? next : prev;
-        });
-      }
-    };
-    window.addEventListener('message', handler);
-    return () => window.removeEventListener('message', handler);
-  }, []);
-
-  React.useEffect(() => {
-    iframeRef.current?.contentWindow?.postMessage({ type: 'theme', dark: isDark }, '*');
-  }, [isDark]);
-
   return (
-    <iframe
-      ref={iframeRef}
-      srcDoc={buildSrcdoc(html, null, isDark)}
-      sandbox="allow-scripts"
-      style={{ width: '100%', height, border: 'none' }}
-      title="HTML output preview"
-    />
+    trimmed.startsWith('<!doctype html') ||
+    trimmed.startsWith('<html>') ||
+    trimmed.startsWith('<html ')
   );
 }
 
@@ -109,7 +74,7 @@ function ResultValue({ value }: { value: unknown }) {
       );
     }
     if (isHtmlString(value)) {
-      return <HtmlPreview html={value} />;
+      return <SandboxedHtmlIframe html={value} title="HTML output preview" />;
     }
     return <span className="text-sm font-mono break-all">{value}</span>;
   }
